docs: added package info and missing javadoc
