+++ src/main/java/osz/imt/mts/mts11umfrage/controller/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This Package contains all controller needed for wiring the backend to the front end. As well as
+ * the required constants.
+ */
+
+package osz.imt.mts.mts11umfrage.controller;
