+++ src/main/java/osz/imt/mts/mts11umfrage/controller/utils/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Package containing utility classes respective for controller classes.
+ *
+ * <p>Created by: Jack</p>
+ * <p>Date: 11.01.2023</p>
+ */
+
+package osz.imt.mts.mts11umfrage.controller.utils;
