--- src/main/java/osz/imt/mts/mts11umfrage/data/QuestionTypes.java
@@ -1,31 +1,0 @@
-package osz.imt.mts.mts11umfrage.data;
-
-/**
- * Question Type Enumeration.
- *
- * <p>Created by: Jack</p>
- * <p>Date: 10.10.2022</p>
- */
-public enum QuestionTypes {
-  /**
-   * Boolean value answer.
-   */
-  BOOLEAN,
-  /**
-   * Single answer is possible.
-   */
-  SINGLEANSWER,
-  /**
-   * A Multiple choice question. Multiple answers can be selected.
-   */
-  MULTIPLECHOICE,
-  /**
-   * An Input question answer.
-   */
-  INPUT,
-  /**
-   * Combo fields.
-   */
-  COMBO
-
-}
