+++ src/main/java/osz/imt/mts/mts11umfrage/dto/package-info.java
@@ -1,0 +1,11 @@
+/**
+ * Package contains all Data Transfer Objects required to provide a secure communication between the
+ * front and backend. Contracts for Version one and Version two of the Json Responses exist.
+ *
+ * <p>Created by: Jack</p>
+ * <p>Date: 11.01.2023</p>
+ *
+ * @author Jacek Langer
+ */
+
+package osz.imt.mts.mts11umfrage.dto;
