+++ src/main/java/osz/imt/mts/mts11umfrage/dto/util/QuestionTypes.java
@@ -1,0 +1,31 @@
+package osz.imt.mts.mts11umfrage.dto.util;
+
+/**
+ * Question Type Enumeration.
+ *
+ * <p>Created by: Jack</p>
+ * <p>Date: 10.10.2022</p>
+ */
+public enum QuestionTypes {
+  /**
+   * Boolean value answer.
+   */
+  BOOLEAN,
+  /**
+   * Single answer is possible.
+   */
+  SINGLEANSWER,
+  /**
+   * A Multiple choice question. Multiple answers can be selected.
+   */
+  MULTIPLECHOICE,
+  /**
+   * An Input question answer.
+   */
+  INPUT,
+  /**
+   * Combo fields.
+   */
+  COMBO
+
+}
