+++ src/main/java/osz/imt/mts/mts11umfrage/dto/util/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Package for constant data used in application.
+ *
+ * <p>Created by: Jack</p>
+ * <p>Date: 11.01.2023</p>
+ */
+
+package osz.imt.mts.mts11umfrage.dto.util;
