+++ src/main/java/osz/imt/mts/mts11umfrage/dto/v1/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Package containing the dtos for the version 1 of the json response.
+ *
+ * <p>Created by: Jack</p>
+ * <p>Date: 11.01.2023</p>
+ */
+
+package osz.imt.mts.mts11umfrage.dto.v1;
