+++ src/main/java/osz/imt/mts/mts11umfrage/dto/v2/JsonResponseEvaluationV2Dto.java
@@ -8,11 +8,20 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import osz.imt.mts.mts11umfrage.dto.Evaluation;
 
+
 /**
- * .
+ * Response object for Json response of version 2.
+ *
+ * <p>Example response</p>
+ * <p>
+ * {@code { "questionid":1, "question":"QuestionText", "answerOptions":["answer",...],
+ * "userAnswers":{ "userSessionId": "id", "userAnswers": ["answer"] } } }
+ * </p>
  *
  * <p>Created by: Jack</p>
  * <p>Date: 07.12.2022</p>
+ *
+ * @author Jacek Langer
  */
 @Builder
 @Getter
