+++ src/main/java/osz/imt/mts/mts11umfrage/dto/v2/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Package containing the version 2 response dto.
+ *
+ * <p>Created by: Jack</p>
+ * <p>Date: 11.01.2023</p>
+ */
+
+package osz.imt.mts.mts11umfrage.dto.v2;
