+++ src/main/java/osz/imt/mts/mts11umfrage/models/package-info.java
@@ -1,8 +1,10 @@
 /**
- * Package info.
+ * Package containing all tables managed by Hibernate.
  *
  * <p>Created by: Jack</p>
  * <p>Date: 20.10.2022</p>
+ *
+ * @author Jacek Langer
  */
 
 @org.hibernate.annotations.GenericGenerator(
