+++ src/main/java/osz/imt/mts/mts11umfrage/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Package combining the logic of the application. This is the main package of the application.
+ *
+ * <p>Created by: Jack</p>
+ * <p>Date: 11.01.2023</p>
+ */
+
+package osz.imt.mts.mts11umfrage;
