+++ src/main/java/osz/imt/mts/mts11umfrage/pythonHandler/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Package for the python wrapper .
+ *
+ * <p>Created by: Jack</p>
+ * <p>Date: 11.01.2023</p>
+ */
+
+package osz.imt.mts.mts11umfrage.pythonHandler;
