+++ src/main/java/osz/imt/mts/mts11umfrage/repository/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Package containing all Repositories used to access and manage respective tables and data.
+ *
+ * <p>Created by: Jack</p>
+ * <p>Date: 11.01.2023</p>
+ */
+
+package osz.imt.mts.mts11umfrage.repository;
