+++ src/main/java/osz/imt/mts/mts11umfrage/service/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Package for services. .
+ *
+ * <p>Created by: Jack</p>
+ * <p>Date: 11.01.2023</p>
+ */
+
+package osz.imt.mts.mts11umfrage.service;
