+++ src/main/java/osz/imt/mts/mts11umfrage/utils/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Package for non sepecific utilities.
+ *
+ * <p>Created by: Jack</p>
+ * <p>Date: 11.01.2023</p>
+ */
+
+package osz.imt.mts.mts11umfrage.utils;
